search 页面类别菜单独立为组件
