+++ src/Page/Search/Component/TypeMenu/index.ts
@@ -1,0 +1,1 @@
+export {default} from './Container';
